[FE] refactor : 페이지컴포넌트 이름 변경
- 상수 네이밍에 맞게 변경했음
